refactor: extract course staff authorization tests
