+++ cms/djangoapps/contentstore/rest_api/v1/views/tests/test_xblock.py
@@ -1,64 +1,24 @@
-import os
-import tarfile
-import tempfile
 from unittest.mock import patch
 from django.http import JsonResponse
 
 from django.urls import reverse
-from path import Path as path
 from rest_framework import status
-from rest_framework.test import APITestCase, APIRequestFactory
+from rest_framework.test import APITestCase
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
-from xmodule.modulestore.tests.factories import SampleCourseFactory, CourseFactory
-from opaque_keys.edx.keys import CourseKey
 
-from cms.djangoapps.contentstore.tests.utils import CoursesWithStaffMixin
-from cms.djangoapps.contentstore.rest_api.v1.views.xblock import toggles, handle_xblock
-from common.djangoapps.student.tests.factories import UserFactory, GlobalStaffFactory, InstructorFactory
+from cms.djangoapps.contentstore.tests.test_utils import AuthorizeStaffTestCase
 
 
-class CourseAPITestcase(CoursesWithStaffMixin):
-    """ setup for studio content API tests """
-
-    def tearDown(self):
-        super().tearDown()
-        self.client.logout()
-
-    @classmethod
-    def create_course_from_course_key(cls, course_key):
-        return CourseFactory.create(
-            org=course_key.org,
-            course=course_key.course,
-            run=course_key.run
-        )
-
-    def make_request(self, course_id=None, data=None):
-        raise NotImplementedError
-
-    def get_url(self, course_key):
-        return reverse('cms.djangoapps.contentstore:v1:studio_content', kwargs={'course_id': course_key})
-
-    def test_403_if_student(self):
-        self.client.login(username=self.student.username, password=self.password)
-        response = self.make_request()
-        assert response.status_code == status.HTTP_403_FORBIDDEN
-
-    def test_403_if_instructor_in_another_course(self):
-        self.client.login(
-            username=self.other_course_instructor.username,
-            password=self.password
-        )
-        response = self.make_request()
-        assert response.status_code == status.HTTP_403_FORBIDDEN
-
-
-class XblockViewPostTest(CourseAPITestcase, ModuleStoreTestCase, APITestCase):
+class XblockViewPostTest(AuthorizeStaffTestCase, ModuleStoreTestCase, APITestCase):
     """
     Test CRUD operations on xblocks
     """
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
+
+    def get_url(self, course_key):
+        return reverse('cms.djangoapps.contentstore:v1:studio_content', kwargs={'course_id': course_key})
 
     def get_test_data(self, course_id):
         return {
@@ -106,32 +66,7 @@ class XblockViewPostTest(CourseAPITestcase, ModuleStoreTestCase, APITestCase):
         response = self.client.post(url)
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
-    def test_200_global_staff(self):
-        self.client.login(username=self.global_staff.username, password=self.password)
-        response = self.make_request()
-        assert response.status_code == status.HTTP_200_OK
-
-    def test_200_course_instructor(self):
-        self.client.login(username=self.course_instructor.username, password=self.password)
-        response = self.make_request()
-        assert response.status_code == status.HTTP_200_OK
-
     def test_xblock_handler_called_with_correct_arguments(self):
         self.client.login(username=self.course_instructor.username, password=self.password)
         response = self.make_request(assert_xblock_handler_call=True)
         assert response.status_code == status.HTTP_200_OK
-
-    # a test case for the create operation
-    # def test_create_an_xblock(self, mock_use_studio_content_api, mock_handle_xblock):
-    #     key = self.course_key.html_id()
-    #     # create an xblock
-    #     url = reverse('cms.djangoapps.contentstore:v1:studio_content', args=[key])
-
-    #     # send a post request
-    #     response = self.client.post(url, data={
-    #         "parent_locator": key,
-    #         "category": "html",
-    #         "courseKey": key,
-    #     }, format='json')
-
-    #     self.assertEqual(response.status_code, status.HTTP_200_OK)
